doubles -> floats, multi print -> helper

+++ Payment.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class Payment extends Transaction {
     static int currentPaymentNumber;
 
-    public Payment(Client client, Date paymentDate, double amount) {
+    public Payment(Client client, Date paymentDate, float amount) {
         this.ID = ++currentPaymentNumber;
         this.client = client;
         this.transactionAmount = amount;
