+++ RepairPrice.java
@@ -6,10 +6,10 @@ package app;
 public class RepairPrice {
     String brand;
     String tier;
-    double price;
+    float price;
     int daysRequired;
 
-    RepairPrice(String brand, String tier, double price, int daysRequired) {
+    RepairPrice(String brand, String tier, float price, int daysRequired) {
         this.brand = brand;
         this.tier = tier;
         this.price = price;
